requirements.txt
keboola.component==1.4.4
keboola.utils==1.1.0
langchain==0.3.2
llama_index==0.11.16
numpy>=1.21,<2
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.39.0
lancedb==0.6.13 
